Finish

+++ pkg/models/product.go
@@ -4,6 +4,6 @@ type Product struct {
 	Code    int64   `json:"good_code"`
 	Name    string  `json:"good_name"`
 	Size    float64 `json:"good_size"`
-	Value   int64   `json:"good_value"`
+	Value   int     `json:"good_value"`
 	StockId int     `json:"warehouse_id"`
 }
